haven/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pymongo==4.10.1
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.27.2
requests==2.32.3
pypdf==5.1.0
python-multipart==0.0.18
